+++ internal/service/server/server.go
@@ -1,0 +1,11 @@
+package server
+
+import (
+	pb "gitlab.com/gitlab-org/gitaly-proto/go"
+)
+
+type server struct{}
+
+func NewServer() pb.ServerServiceServer {
+	return &server{}
+}